+++ specification/_types/Knn.ts
@@ -18,7 +18,7 @@
  */
 
 import { Field } from '@_types/common'
-import { long, double, float } from '@_types/Numeric'
+import { long, float } from '@_types/Numeric'
 import { QueryContainer } from './query_dsl/abstractions'
 
 export type QueryVector = float[]
@@ -39,7 +39,7 @@ export interface KnnQuery {
   /** Filters for the kNN search query */
   filter?: QueryContainer | QueryContainer[]
   /** The minimum similarity for a vector to be considered a match */
-  similarity?: double
+  similarity?: float
 }
 
 /** @variants container */
